testlab: move supertest utilities to new file
